Revert "Merge branch 'charm' of charmgit:charm into charm"
This reverts commit 0cebe4608c328606b7fc33a52a29d4a5ad7229bd, reversing
changes made to 4764e0bf186b8bbf48607414d0f5fd8fe2cf50bb.

+++ src/conv-core/converse.h
@@ -1453,7 +1453,7 @@ void CthUserAddListeners(CthThread th);
 
 /****** Isomalloc: Migratable Memory Allocation ********/
 /*Simple block-by-block interface:*/
-void *CmiIsomalloc(size_t sizeInBytes, CthThread tid);
+void *CmiIsomalloc(int sizeInBytes, CthThread tid);
 void *CmiIsomallocAlign(size_t align, size_t size, CthThread t);
 void  CmiIsomallocPup(pup_er p,void **block);
 void  CmiIsomallocFree(void *block);
