+++ src/util/pup_stl.h
@@ -85,7 +85,7 @@ inline void operator|(PUP::er &p,std::complex<T> &v)
 template <class charType> 
 inline void operator|(PUP::er &p,typename std::basic_string<charType> &v)
 {
-  size_t nChar=v.length();
+  int nChar=v.length();
   p|nChar;
   if (p.isUnpacking()) { //Unpack to temporary buffer
     charType *buf=new charType[nChar];
@@ -101,7 +101,7 @@ inline void operator|(PUP::er &p,typename std::basic_string<charType> &v)
 inline void operator|(PUP::er &p,std::string &v)
 {
   p.syncComment(PUP::sync_begin_object,"std::string");
-  size_t nChar=v.length();
+  int nChar=v.length();
   p|nChar;
   if (p.isUnpacking()) { //Unpack to temporary buffer
     char *buf=new char[nChar];
@@ -120,8 +120,8 @@ inline void operator|(PUP::er &p,std::string &v)
 
 //Impl. util: pup the length of a container
 template <class container>
-inline size_t PUP_stl_container_size(PUP::er &p,container &c) {
-  size_t nElem=c.size();
+inline int PUP_stl_container_size(PUP::er &p,container &c) {
+  int nElem=c.size();
   p|nElem;
   return nElem; 
 }
@@ -141,11 +141,11 @@ inline void PUP_stl_container_items(PUP::er &p,container &c) {
 template <class container,class dtype>
 inline void PUP_stl_container(PUP::er &p,container &c) {
   p.syncComment(PUP::sync_begin_array);
-  size_t nElem=PUP_stl_container_size(p,c);
+  int nElem=PUP_stl_container_size(p,c);
   if (p.isUnpacking()) 
   { //Unpacking: Extract each element and push_back:
     c.resize(0);
-    for (size_t i=0;i<nElem;i++) {
+    for (int i=0;i<nElem;i++) {
       p.syncComment(PUP::sync_item);
       dtype n;
       p|n;
@@ -160,10 +160,10 @@ inline void PUP_stl_container(PUP::er &p,container &c) {
 template <class container,class dtype>
 inline void PUP_stl_map(PUP::er &p,container &c) {
   p.syncComment(PUP::sync_begin_list);
-  size_t nElem=PUP_stl_container_size(p,c);
+  int nElem=PUP_stl_container_size(p,c);
   if (p.isUnpacking()) 
   { //Unpacking: Extract each element and insert:
-    for (size_t i=0;i<nElem;i++) {
+    for (int i=0;i<nElem;i++) {
       dtype n;
       p|n;
       c.insert(n);
